+++ javascript/main.js
@@ -1,14 +1,16 @@
 function api_post(path, payload, callback) {
     var xhr = new XMLHttpRequest();
-    xhr.open("POST", path, false);
-    xhr.onreadystatechange = function() { if (xhr.readyState === 4 && xhr.status === 200) { callback(JSON.parse(xhr.responseText)); } }
+    xhr.open("POST", path, true);
+    xhr.onload = function() { callback(JSON.parse(xhr.responseText)); };
     xhr.setRequestHeader("Content-type", "application/json");
     xhr.send(JSON.stringify(payload));
 }
 
 function when_loaded() {
     api_post("/style-editor/check-api/", {}, function(x) { console.log( "Style Editor Check API", x['value'] )});
-    document.getElementById('style_editor_grid').addEventListener('keydown', function(event){
+    globalThis.selectedRows = [];
+    grid = document.getElementById('style_editor_grid');
+    grid.addEventListener('keydown', function(event){
         if (event.ctrlKey === true) {
             event.stopImmediatePropagation();
             span = event.target.querySelector("span");
@@ -30,28 +32,28 @@ function when_loaded() {
 
         // if backspace or delete are pressed, and we're over the selected row, delete it
         if (event.key === "Backspace" || event.key === "Delete") { 
-            if (globalThis.selectedRow) { 
-                {
-                    api_post("/style-editor/delete-style", 
-                         {"value":row_style_name(globalThis.selectedRow)}, 
-                         function(x){document.getElementById("style_editor_handle_api").click()} );
-                }
-            }
+            globalThis.selectedRows.forEach( function(row) { 
+                api_post("/style-editor/delete-style", 
+                        {"value":row_style_name(row)}, 
+                        function(x){} );    
+            });
+            document.getElementById("style_editor_handle_api").click();
+            globalThis.selectedRows = [];
         } 
 
         // if we get to here, stop the keypress from propogating
         event.stopImmediatePropagation(); 
     }, { capture: true });
 
-    document.getElementById('style_editor_grid').addEventListener('contextmenu', function(event){
+    grid.addEventListener('contextmenu', function(event){
         if(event.shiftKey) { return; }
-        unselect_row();    
+        if(!event.ctrlKey) { unselect_rows(); }
         row = event.target.closest("tr");
         if (row) { select_row(row); event.stopImmediatePropagation(); event.preventDefault(); }  
     }, { capture: true });
 
-    document.getElementById('style_editor_grid').addEventListener('click', function(event){
-        unselect_row()
+    grid.addEventListener('click', function(event){
+        unselect_rows()
     }, { capture: true });
 }
 
@@ -60,16 +62,13 @@ function row_style_name(row) {
 }
 
 function select_row(row) {
-    globalThis.savedStyle = row.style;
-    globalThis.selectedRow = row;
+    globalThis.selectedRows.push(row);
     row.style.backgroundColor = "#840";
 }
 
-function unselect_row() {
-    if (globalThis.selectedRow) {
-        globalThis.selectedRow.style = globalThis.savedStyle;
-        globalThis.selectedRow = null;
-    }
+function unselect_rows() {
+    globalThis.selectedRows.forEach( function(row){ row.style.backgroundColor = '';  })
+    globalThis.selectedRows = []
 }
 
 function press_refresh_button(tab) {